add xctesthtmlreport

+++ fastlane/actions/xcodegen.rb
@@ -2,17 +2,21 @@ module Fastlane
   module Actions
     class XcodegenAction < Action
       def self.run(params)
-        Actions.sh("which xcodegen", log: false, error_callback: proc do |error_output|
+        other_action.setup_local_bin
+
+        binary_path = File.join(lane_context[SharedValues::LOCAL_BINARY_PATH], "xcodegen")
+
+        unless File.exist?(binary_path)
           UI.message("Installing XcodeGen...")
 
           Actions.sh("curl -OL https://github.com/yonaskolb/XcodeGen/releases/download/2.25.0/xcodegen.zip")
           Actions.sh("unzip -q -o xcodegen.zip")
           Actions.sh("rm xcodegen.zip")
-          Actions.sh("xcodegen/install.sh")
+          Actions.sh("mv xcodegen/bin/xcodegen #{binary_path.shellescape}")
           Actions.sh("rm -rf xcodegen")
-        end)
+        end
 
-        Actions.sh("/usr/local/bin/xcodegen")
+        Actions.sh("#{binary_path.shellescape}")
       end
 
       def self.description
